pyproject.toml
[build-system]
build-backend = "mesonpy"
requires = ["meson-python>=0.15.0"]

[project]
name = "wisio"
version = "0.0.1"
authors = [{ name = "Izzet Yildirim", email = "izzetcyildirim@gmail.com" }]
maintainers = [{ name = "Izzet Yildirim", email = "izzetcyildirim@gmail.com" }]
requires-python = ">=3.8"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Build Tools",
    "Typing :: Typed",
]
dependencies = [
    "dask~=2023.9.0",
    "dask_jobqueue==0.8.2",
    "distributed~=2023.9.0",
    "hydra-core~=1.3.0",
    "inflect==7.0",
    "jinja2>=3.0",
    "numpy==1.24.3",
    "matplotlib>=3.2.1",
    "pandas>=2.1",
    "pyarrow>=13",
    "pyyaml>=5.4",
    "rich==13.6.0",
    "scikit-learn>=1.4",
    "scipy>=1.10",
    "venn==0.1.3",
]

[project.optional-dependencies]
darshan = ["darshan>=3.4"]

[project.scripts]
wisio = "wisio.__main__:main"

[project.urls]
documentation = "https://wisio.readthedocs.io"
source = "https://github.com/izzet/wisio"
tracker = "https://github.com/izzet/wisio/issues"

[tool.ruff.format]
quote-style = "preserve"

[tool.ruff.lint.pycodestyle]
max-line-length = 100

[tool.spin]
package = 'wisio'
